feat: create separate cue caches for different registry configs
While we SHOULD assume CUE modules *should* be immutable and never change, varying
registry configurations may change what modules are actually pulled.

Additionally, we should avoid modules being present in the cache that were pulled
from a registry that is no longer included in the current registry list. This may
allow a bundle to template without issues on one machine, while failing on another,
even with a matching configuration.

This change takes a hash of the sorted registry list and appends that to the CUE
cache directory used by Odin. Identical configurations should always behave the same
this way, assuming the contents of the underlying registries do not change.

Signed-off-by: Stefan Nuxoll <stefan@nuxoll.me>